split prediction from feature extraction
